requirements.txt
Django==2.1.5
django-rest-auth==0.9.3
djangorestframework==3.9.1
